Removed unnecessary cast to float

+++ model.py
@@ -32,4 +32,4 @@ class Adversary(nn.Module):
         s = torch.sigmoid((1 + torch.abs(self.c)) * logits)
         z_hat = torch.cat((s, s*targets, s*(1-targets)),dim=1) @ self.w2 + self.b
 
-        return z_hat, torch.sigmoid(z_hat).float()
+        return z_hat, torch.sigmoid(z_hat)
